+++ R/import_data_module.R
@@ -833,20 +833,25 @@ import_data_server <- function(id,
         <ul>
           <li> <strong>Cross-sectional Population Data (Pop Data)</strong>
                   <p>A dataset with one row per sample and columns for
-                  antigen isotype, quantitative antibody results, participant ID, and age
+                  antigen isotype, quantitative antibody results,
+                  participant ID, and age
                   in years.</p>
         <ul>
               <li>Additional columns and variables can be included for
               stratification</li>
               <li>Age unit is years, decimal points are fine</li>
               <li>The scale of the antibody response variable must be the
-              same as the longitudinal antibody decay data (seroresponse parameters) </li>
+              same as the longitudinal antibody decay data
+              (seroresponse parameters) </li>
               <li>Do not upload any identifying health information </li>
-              <li>The antigen-isotype variable must be named <code>antigen_iso</code> </li>
+              <li>The antigen-isotype variable must be named
+              <code>antigen_iso</code> </li>
    </p>
             </ul>
-          <li><strong> Longitudinal Antibody Seroresponse Parameters (Seroresponse Data) </strong></li>
-          <p>A data set containing within-host antibody seroresponse parameters fit
+          <li><strong> Longitudinal Antibody Seroresponse Parameters
+          (Seroresponse Data) </strong></li>
+          <p>A data set containing within-host antibody seroresponse
+          parameters fit
           using a two-phase within-host Bayesian hierarchical framework
           obtaining predictive posterior samples using Markov chain
           Monte Carlo sampling. Note that variable names <u>must</u>
@@ -860,14 +865,17 @@ import_data_server <- function(id,
               <li><code>antigen_iso</code>: antigen-isotype</li>
               <li><code>y0</code>: baseline antibody level</li>
               <li><code>y1</code>: antibody peak level </li>
-              <li><code>t1</code>: time from symptom onset to peak antibody response
+              <li><code>t1</code>: time from symptom onset to peak
+              antibody response
               (in days) </li>
-              <li><code>alpha</code>: antibody decay rate (1/days for the current
+              <li><code>alpha</code>: antibody decay rate (1/days for the
+              current
               longitudinal parameter sets)</li>
               <li><code>r</code>: shape factor of antibody decay</li>
             </ul>
         </p>
-           <li> <strong> Antibody cross reactivity, measurement noise and limits of detection (Noise Data)</strong>
+           <li> <strong> Antibody cross reactivity, measurement noise
+           and limits of detection (Noise Data)</strong>
                   <p>A dataset containing the following variables,
                   specifying noise parameters for each antigen isotype.
                   Note that variable names <u>must</u> follow these guidelines.
@@ -875,27 +883,36 @@ import_data_server <- function(id,
                   <a hfref=https://onlinelibrary.wiley.com/doi/10.1002/sim.8578>
                   Teunis (2020)</a>.
                   <ul>
-                    <li><code>antigen_iso</code>: antigen isotype whose noise parameters are
-                    being specified on each row</li>
-                    <li><code>y.low</code>: Lower limit of detection of the antibody assay
+                    <li><code>antigen_iso</code>: antigen isotype whose noise
+                    parameters are being specified on each row</li>
+                    <li><code>y.low</code>: Lower limit of detection of the
+                    antibody assay
                     </li>
-                    <li><code>y.high</code>: Upper limit of detection of the antibody assay
+                    <li><code>y.high</code>: Upper limit of detection of the
+                    antibody assay
                     </li>
                     <li><code>nu</code>: biological noise</li>
                        <ul>
                             <li>
-                              <em> Typically comes from cross-reactivity with other molecules or other pathogens.
-                              Biological noise is defined as the 95th percentile of the distribution of antibody
-                              responses to the antigen-isotype in a population with no exposure.</em>
+                              <em> Typically comes from cross-reactivity with
+                              other molecules or other pathogens.
+                              Biological noise is defined as the 95th percentile
+                              of the distribution of antibody
+                              responses to the antigen-isotype in a population
+                              with no exposure.</em>
                             </li>
                           </ul>
                     <li><code>eps</code>: measurement noise</li>
                        <ul>
                             <li>
-                              <em>Represents measurement error from the laboratory testing process
-                              It is defined by a CV (coefficient of variation) as the ratio of
-                              the standard deviation to the mean for replicates. Note that the CV should ideally be
-                              measured across plates rather than within the same plate.</em>
+                              <em>Represents measurement error from the
+                              laboratory testing process
+                              It is defined by a CV (coefficient of variation)
+                              as the ratio of
+                              the standard deviation to the mean for replicates.
+                              Note that the CV should ideally be
+                              measured across plates rather than within the
+                              same plate.</em>
                             </li>
                           </ul>